+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - add unit converter const factories
 - computations with units
 - add units in quantity classes as static variables
+- rename Duration quantity to DurationQuantity to avoid collision with core Duration class
 
 ## 0.0.1